refactor: alertmodal 공통 컴포넌트로 분리 및 검색필드 common 폴더로 이동
